renamed the data model files
