+++ lib/common/rest_api/file_helper.dart
@@ -36,8 +36,8 @@ import 'package:communitypod/common/rest_api/operations.dart';
 import 'package:communitypod/constants/app.dart';
 import 'package:communitypod/constants/paths.dart';
 import 'package:communitypod/constants/turtle_structures.dart';
-import 'package:communitypod/models/note.dart';
-import 'package:communitypod/models/note_content.dart';
+import 'package:communitypod/models/news.dart';
+import 'package:communitypod/models/news_content.dart';
 import 'package:communitypod/notes/list_my_notes_screen.dart';
 import 'package:communitypod/notes/view_note.dart';
 import 'package:communitypod/utils/encryption.dart';
