+++ lib/models/news_content.dart
