+++ lib/models/own_news.dart
@@ -28,7 +28,7 @@ library;
 import 'package:solidpod/solidpod.dart';
 
 import 'package:communitypod/constants/turtle_structures.dart';
-import 'package:communitypod/models/note_content.dart';
+import 'package:communitypod/models/news_content.dart';
 
 /// Data model for user's news file
