+++ lib/models/selected_news.dart
