+++ lib/notes/edit_note.dart
@@ -31,7 +31,7 @@ import 'package:flutter/services.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
 import 'package:solidui/solidui.dart';
 
-import 'package:communitypod/models/note.dart';
+import 'package:communitypod/models/news.dart';
 import 'package:communitypod/notes/view_note.dart';
 import 'package:communitypod/widgets/note_edit_scroll_view.dart';
