+++ lib/utils/turtle/note_serializer.dart
@@ -28,7 +28,7 @@ library;
 import 'package:rdflib/rdflib.dart';
 
 import 'package:communitypod/constants/turtle_structures.dart';
-import 'package:communitypod/models/note_content.dart';
+import 'package:communitypod/models/news_content.dart';
 import 'package:communitypod/utils/encryption.dart';
 import 'package:communitypod/utils/turtle/parsing_utils.dart';
