+++ lib/widgets/note_list_del_button.dart
@@ -33,7 +33,7 @@ import 'package:communitypod/common/rest_api/file_helper.dart';
 import 'package:communitypod/constants/app.dart';
 import 'package:communitypod/constants/colours.dart';
 import 'package:communitypod/constants/ui.dart';
-import 'package:communitypod/models/selected_note.dart';
+import 'package:communitypod/models/selected_news.dart';
 import 'package:communitypod/widgets/err_card.dart';
 import 'package:communitypod/widgets/loading_animation.dart' as loading;
