+++ lib/widgets/note_list_revoke_button.dart
@@ -32,7 +32,7 @@ import 'package:solidui/solidui.dart';
 import 'package:communitypod/constants/app.dart';
 import 'package:communitypod/constants/colours.dart';
 import 'package:communitypod/constants/ui.dart';
-import 'package:communitypod/models/note.dart';
+import 'package:communitypod/models/news.dart';
 import 'package:communitypod/widgets/loading_animation.dart' as loading;
 
 /// A revoke button widget for updating the log record for a list
