+++ lib/widgets/note_save_button.dart
@@ -32,7 +32,7 @@ import 'package:solidui/solidui.dart';
 
 import 'package:communitypod/common/rest_api/file_helper.dart';
 import 'package:communitypod/constants/colours.dart';
-import 'package:communitypod/models/note.dart';
+import 'package:communitypod/models/news.dart';
 
 /// A stylised save button widget which on click saves the note content
 /// Pod. External notes are written to the note owner's Pod. News created
